Added automatic reduction of images.
Macro \image reduces the image width to \hsize if it is larger
than \hsize and similarly in the vertical direction. This may not work
for portrait images whose width is larger than \hsize but even after
reducing width the height is still larger than \vsize. Manual
reduction must be specified in such a case.